Implement evaluation script
This commit also includes a change in the environment that allows
it to take in an experiment setting. This will cause the environment
to always fix the setting when resetting which allows us to evaluate
the different settings without depending on the number random generator.
It also includes a change to propagate the item collected after each
call to `step` inside the `info` dictionary.